requirements.txt
numpy>=1.19.0
open3d>=0.13.0
opencv-python>=4.5.0
pillow>=8.0.0
torch>=1.9.0
transformers>=4.15.0